% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta11.r
\name{ComputeConditionalDensity}
\alias{ComputeConditionalDensity}
\title{Calculeaza densitatea conditionala}
\usage{
ComputeConditionalDensity(fct, Which = "X")
}
\arguments{
\item{Which}{carei variabile sa i se calculeze densitatea conditionala}

\item{f}{Functia de densitate comuna}
}
\value{
functia densitate marginala
}
\description{
Primeste ca parametru o densitatate comuna a doua variabile returneaza densitatea conditionala
}
\examples{
ComputeConditionalDensity(function (x, y) <- {return (x + y)}, "X")
}
