% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta11.r
\name{ComputeMarginalDistribution}
\alias{ComputeMarginalDistribution}
\title{Calculeaza densitatea marginala}
\usage{
ComputeMarginalDistribution(fct, Which = "X")
}
\arguments{
\item{Which}{carei variabile sa i se calculeze densitatea}

\item{f}{Functia de densitate comuna}
}
\value{
functia densitate marginala
}
\description{
Primeste ca parametru o densitatate comuna a doua variabile returneaza densitatea marginala
}
\examples{
ComputeMarginalDistribution(function (x, y) <- {return (x + y)}, "X")
}
