% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta6.r
\name{ComputeMeanDispersion}
\alias{ComputeMeanDispersion}
\title{Functie care calculeaza media si dispersia}
\usage{
ComputeMeanDispersion(
  g,
  nume = "",
  a = 0,
  b = 0,
  n = 0,
  lambda = 0,
  sigma = 0,
  m = 0
)
}
\arguments{
\item{g}{functie continua de la utilizator (daca nu este continua pot aparea erori)}

\item{nume}{numele repartitiei alese}

\item{a}{parametru optional}

\item{b}{parametru optional}

\item{n}{parametru optional}

\item{lambda}{parametru optinal}

\item{sigma}{parametru optional}

\item{m}{parametru optional}
}
\value{
media si dispersia
}
\description{
Functie care returneaza media si densitatea variabilei g(X), unde X este o repartitie cunoscuta
si g este o functie continua data de utilizator
Apeleaza functia interna ExecuteMeanDispersion pentru a calcula rezultatul necesar
}
\examples{
ComputeMeanDispersion (function(x){return(x^4 + 5)}, function(x){return 1/(1-0.5)}, 1, 0.5)
}
