% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta8.r
\name{DistributionsSynthesis}
\alias{DistributionsSynthesis}
\title{Functie ce ofera informatii despre diverse distributii cunoscute}
\usage{
DistributionsSynthesis(name)
}
\arguments{
\item{name}{Numele repartitiei despre care vrem sa aflam informatii}
}
\value{
un sir de caractere ce afiseaza informatii despre repartitia in cauza
}
\description{
Functia primeste un nume si afiseaza informatii relevante si interesante
despre repartitia in cauza
}
\examples{
DistributionsSynthesis ("Uniform")
}
