% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cerinta6.r
\name{ExecuteMeanDispersion}
\alias{ExecuteMeanDispersion}
\title{Functie interna pentru medie si dispersie}
\usage{
ExecuteMeanDispersion(g, densityFunction, lowerBound, upperBound)
}
\arguments{
\item{g}{functie de la utilizator}

\item{densityFunction}{functie de densitate}

\item{lowerBound}{capatul stanga al integralei}

\item{upperBound}{capatul dreapta al integralei}
}
\value{
rezultatul sub forma de sir de caractere
}
\description{
Functie interna care returneaza media si densitatea variabilei g(X), unde X este o repartitie cunoscuta
si g este o functie continua data de utilizator
Este o functie interna deoarece este apelata de catre functia pentru cerinta 6
}
\examples{
ExecuteMeanDispersion (function(x){return(x^4 + 5)}, function(x){return 1/(1-0.5)}, 1, 0.5)
}
